# specs/013-fastapi-jwt-auth/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Event-AI Authentication API
  version: 1.0.0
  description: |
    JWT-based authentication with OAuth2 password grant for Event-AI platform.
    All endpoints under `/api/v1/auth` require HTTPS in production.

servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.event-ai.com/api/v1
    description: Production

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user account
      description: |
        Creates a new user account with email and password.
        On success, returns user profile, access token, and refresh token.
        Passwords are validated for complexity (12+ chars with special chars).
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          description: Validation error (password too weak, invalid email format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: VALIDATION_PASSWORD_STRENGTH
                  message: Password must be at least 12 characters with uppercase, lowercase, digit, and special character
                  field: password
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: CONFLICT_EMAIL_EXISTS
                  message: An account with this email already exists
                  field: email
        '429':
          description: Rate limit exceeded (10/hour per IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '503':
          description: Service unavailable (database down)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in with OAuth2 password grant
      description: |
        Authenticates user with email and password using OAuth2 password grant.
        Accepts `application/x-www-form-urlencoded` (per OAuth2 spec).
        Returns JWT access token (15min) and refresh token (7 days).
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
              required: [username, password]
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
          headers:
            WWW-Authenticate:
              description: OAuth2 Bearer header (always present on success)
              schema:
                type: string
                example: Bearer
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            WWW-Authenticate:
              schema:
                type: string
                example: Bearer
        '403':
          description: Account locked due to failed attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: AUTH_ACCOUNT_LOCKED
                  message: Account locked due to multiple failed login attempts. Try again in 15 minutes.
        '429':
          description: Rate limit exceeded (5/15min per email+IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token using refresh token
      description: |
        Exchanges a valid refresh token for a new access token and new refresh token.
        Implements refresh token rotation: old refresh token is invalidated.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: Log out and invalidate refresh token
      description: |
        Revokes the current refresh token, preventing it from being used to
        obtain new access tokens. Does not blacklist current access token
        (short-lived; expires naturally).
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/password-reset-request:
    post:
      tags: [Authentication]
      summary: Request a password reset token
      description: |
        Generates a one-time password reset token for the given email.
        In development/testing, returns the token directly. In production,
        would trigger an email.
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset token generated (endpoint returns 200 regardless of email existence to prevent email enumeration - returns token only for registered emails in dev)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetTokenResponse'
        '400':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded (5/hour per email)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/password-reset-confirm:
    post:
      tags: [Authentication]
      summary: Confirm password reset with token
      description: |
        Validates the reset token and updates the user's password.
        On success, invalidates all existing refresh tokens for the user
        (terminates all sessions).
      operationId: confirmPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid/expired/used token or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expired:
                  summary: Token expired
                  value:
                    success: false
                    error:
                      code: VALIDATION_TOKEN_EXPIRED
                      message: Password reset token has expired
                used:
                  summary: Token already used
                  value:
                    success: false
                    error:
                      code: VALIDATION_TOKEN_USED
                      message: Password reset token has already been used
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: |
        Returns the authenticated user's profile. Requires valid access token.
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter `Bearer {access_token}` in the Authorization header.
        Access tokens expire in 15 minutes.

  schemas:
    UserRegister:
      $ref: '#/components/schemas/UserRegister'
    UserRead:
      $ref: '#/components/schemas/UserRead'
    Token:
      $ref: '#/components/schemas/Token'
    TokenData:
      $ref: '#/components/schemas/TokenData'
    RefreshTokenRequest:
      $ref: '#/components/schemas/RefreshTokenRequest'
    PasswordResetRequest:
      $ref: '#/components/schemas/PasswordResetRequest'
    PasswordResetConfirm:
      $ref: '#/components/schemas/PasswordResetConfirm'
    PasswordResetTokenResponse:
      $ref: '#/components/schemas/PasswordResetTokenResponse'
    ErrorResponse:
      $ref: '#/components/schemas/ErrorResponse'
    SuccessResponse:
      $ref: '#/components/schemas/SuccessResponse'
    RateLimitError:
      $ref: '#/components/schemas/RateLimitError'

  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer

tags:
  - name: Authentication
    description: User registration, login, token management, password reset